requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
pandas==2.3.2
openpyxl==3.1.5
passlib==1.7.4